fix: Move azurite_client.py to src/providers
